zltplusai/RosBagPlayer/requirements.txt
catkin-pkg==0.4.23
catkin-pkg-modules==0.4.23
catkin-tools==0.7.1
numpy==1.21.1
Pillow==8.3.1
pyglet==1.5.18
rosbag==1.14.11
roscreate==1.14.9
rosgraph==1.14.11
rosgraph-msgs==1.11.3.post2
roslib==1.14.9
roslz4==1.14.11
rosmake==1.14.9
rosmaster==1.14.11
rosmsg==1.14.11
rospkg==1.3.0
rospy==1.14.11
xlib==0.21

